validate with exceptions

+++ Modules/Identity.Application/AppUser/UserAppService.cs
@@ -19,25 +19,64 @@ public class UserAppService(IUserRepository userRepository) : IUserAppService
 
     public async Task<Users> GetByEmailAsync(string email)
     {
+        var isValidEmail = Users.IsValidEmail(email);
+        if (!isValidEmail)
+        {
+            throw new ArgumentException("Invalid email format.");
+        }
         var user = await _userRepository.GetByEmailAsync(email);
+
+        if (user == null)
+        {
+            throw new KeyNotFoundException($"User with email {email} not found.");
+        }
         return user;
     }
 
     public async Task<Users> GetByIdAsync(Guid id)
     {
+        if (id == Guid.Empty)
+        {
+            throw new ArgumentException("Id cannot be empty.", nameof(id));
+        }
+
         var user = await _userRepository.GetByIdAsync(id);
+        if (user == null)
+        {
+            throw new KeyNotFoundException($"User with Id {id} not found.");
+        }
         return user;
     }
 
     public async Task<Users> InsertAsync(UserInput userInput)
     {
         var user = new Users(userInput.Name, userInput.Email, userInput.Password, userInput.Type);
+        var isValidInput = user.Validate();
+        var hasUser = await _userRepository.GetByEmailAsync(userInput.Email);
+        if (!isValidInput)
+        {
+            throw new ArgumentException("Invalid user input.");
+        }
+
+        if (hasUser != null)
+        {
+            throw new ArgumentException($"User with email {userInput.Email} already exists.");
+        }
         await _userRepository.InsertAsync(user);
         return user;
     }
 
     public async Task<Users> UpdateAsync(Guid id, UserInput userInput)
     {
+        if (id == Guid.Empty)
+        {
+            throw new ArgumentException("Id is required"); // needs better error handling and messages
+        }
+        if (userInput == null)
+        {
+            throw new ArgumentException("User information is required");
+        }
+
         var user = await _userRepository.GetByIdAsync(id);
         user.Update(userInput.Name, userInput.Email, userInput.Password, userInput.Type);
         await _userRepository.UpdateAsync(user);
@@ -46,6 +85,10 @@ public class UserAppService(IUserRepository userRepository) : IUserAppService
 
     public async Task DeleteAsync(Guid id)
     {
+        if (id == Guid.Empty)
+        {
+            throw new ArgumentException("Id is required");
+        }
         await _userRepository.DeleteAsync(id);
     }
